公共字段自动填充业务代码开发

+++ sky-server/src/main/java/com/sky/service/impl/CategoryServiceImpl.java
@@ -2,11 +2,16 @@ package com.sky.service.impl;
 
 import com.github.pagehelper.Page;
 import com.github.pagehelper.PageHelper;
+import com.sky.constant.MessageConstant;
+import com.sky.constant.StatusConstant;
 import com.sky.context.BaseContext;
 import com.sky.dto.CategoryDTO;
 import com.sky.dto.CategoryPageQueryDTO;
 import com.sky.entity.Category;
+import com.sky.exception.DeletionNotAllowedException;
 import com.sky.mapper.CategoryMapper;
+import com.sky.mapper.DishMapper;
+import com.sky.mapper.SetmealMapper;
 import com.sky.result.PageResult;
 import com.sky.service.CategoryService;
 import org.springframework.beans.BeanUtils;
@@ -21,11 +26,18 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Autowired
     private CategoryMapper categoryMapper;
+
+    @Autowired
+    private DishMapper dishMapper;
+
+    @Autowired
+    private SetmealMapper setmealMapper;
+
     @Override
     public void update(CategoryDTO categoryDTO) {
         Category category = new Category();
-        category.setUpdateUser(BaseContext.getCurrentId());
-        category.setUpdateTime(LocalDateTime.now());
+//        category.setUpdateUser(BaseContext.getCurrentId());
+//        category.setUpdateTime(LocalDateTime.now());
 
         BeanUtils.copyProperties(categoryDTO, category);
 
@@ -58,8 +70,8 @@ public class CategoryServiceImpl implements CategoryService {
         Category category = Category.builder()
                         .id(id)
                         .status(status)
-                        .updateUser(BaseContext.getCurrentId())
-                        .updateTime(LocalDateTime.now())
+//                        .updateUser(BaseContext.getCurrentId())
+//                        .updateTime(LocalDateTime.now())
                         .build();
         categoryMapper.update(category);
     }
@@ -75,15 +87,15 @@ public class CategoryServiceImpl implements CategoryService {
         Category category = new Category();
         BeanUtils.copyProperties(categoryDTO, category);
 
-        Long createUser = BaseContext.getCurrentId();
-        category.setCreateUser(createUser);
-        category.setUpdateUser(createUser);
+//        Long createUser = BaseContext.getCurrentId();
+//        category.setCreateUser(createUser);
+//        category.setUpdateUser(createUser);
+//
+//        LocalDateTime now = LocalDateTime.now();
+//        category.setCreateTime(now);
+//        category.setUpdateTime(now);
 
-        LocalDateTime now = LocalDateTime.now();
-        category.setCreateTime(now);
-        category.setUpdateTime(now);
-
-        category.setStatus(0);
+        category.setStatus(StatusConstant.DISABLE);
 
         categoryMapper.add(category);
     }
@@ -94,6 +106,15 @@ public class CategoryServiceImpl implements CategoryService {
      */
     @Override
     public void delete(Long id) {
+        Integer count = dishMapper.countByCategoryId(id);
+        if(count > 0) {
+            throw new DeletionNotAllowedException(MessageConstant.CATEGORY_BE_RELATED_BY_DISH);
+        }
+
+        count = setmealMapper.countByCategoryId(id);
+        if(count > 0) {
+            throw new DeletionNotAllowedException(MessageConstant.CATEGORY_BE_RELATED_BY_SETMEAL);
+        }
         categoryMapper.delete(id);
     }
